moved zcashd launch fn to be an associated fn for Zcashd
