+++ tests/integration.rs
@@ -1,9 +1,7 @@
-use zcash_local_net::{launch, network::ActivationHeights};
-
 #[test]
 fn launch_zcashd() {
     tracing_subscriber::fmt().init();
 
-    let zcashd = launch::zcashd(None, None, None, &ActivationHeights::default(), None).unwrap();
+    let zcashd = zcash_local_net::Zcashd::default();
     zcashd.print_stdout();
 }
